+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\alex_\\projects\\pokemon-game\\src\\App.js":"1","C:\\Users\\alex_\\projects\\pokemon-game\\src\\index.js":"2","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\HeaderBlock\\index.js":"3","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Header\\index.js":"4","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Footer\\index.js":"5","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Layout\\index.js":"6","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\PokemonCard\\index.js":"7"},{"size":3691,"mtime":1612097635461,"results":"8","hashOfConfig":"9"},{"size":214,"mtime":1611853674117,"results":"10","hashOfConfig":"9"},{"size":218,"mtime":1611854355061,"results":"11","hashOfConfig":"9"},{"size":293,"mtime":1612029228796,"results":"12","hashOfConfig":"9"},{"size":247,"mtime":1612029272596,"results":"13","hashOfConfig":"9"},{"size":675,"mtime":1612030164189,"results":"14","hashOfConfig":"9"},{"size":1630,"mtime":1612282767767,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"4em567",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\alex_\\projects\\pokemon-game\\src\\App.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\HeaderBlock\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Header\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Footer\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Layout\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\PokemonCard\\index.js",[]]
+[{"C:\\Users\\alex_\\projects\\pokemon-game\\src\\App.js":"1","C:\\Users\\alex_\\projects\\pokemon-game\\src\\index.js":"2","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Header\\index.js":"3","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Footer\\index.js":"4","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Layout\\index.js":"5","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\PokemonCard\\index.js":"6","C:\\Users\\alex_\\projects\\pokemon-game\\src\\routes\\Home\\index.js":"7","C:\\Users\\alex_\\projects\\pokemon-game\\src\\routes\\Game\\index.js":"8","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\MenuHeader\\Menu\\index.js":"9","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\MenuHeader\\index.js":"10","C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\MenuHeader\\NavBar\\index.js":"11"},{"size":458,"mtime":1612290897383,"results":"12","hashOfConfig":"13"},{"size":214,"mtime":1611853674117,"results":"14","hashOfConfig":"13"},{"size":442,"mtime":1612291114187,"results":"15","hashOfConfig":"13"},{"size":247,"mtime":1612029272596,"results":"16","hashOfConfig":"13"},{"size":609,"mtime":1612291467353,"results":"17","hashOfConfig":"13"},{"size":1586,"mtime":1612291928037,"results":"18","hashOfConfig":"13"},{"size":3957,"mtime":1612294198290,"results":"19","hashOfConfig":"13"},{"size":84,"mtime":1612290865787,"results":"20","hashOfConfig":"13"},{"size":541,"mtime":1612295165690,"results":"21","hashOfConfig":"13"},{"size":386,"mtime":1612294398929,"results":"22","hashOfConfig":"13"},{"size":383,"mtime":1612294769433,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"4em567",{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\alex_\\projects\\pokemon-game\\src\\App.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Header\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Footer\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\Layout\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\PokemonCard\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\routes\\Home\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\routes\\Game\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\MenuHeader\\Menu\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\MenuHeader\\index.js",[],"C:\\Users\\alex_\\projects\\pokemon-game\\src\\components\\MenuHeader\\NavBar\\index.js",["46"],{"ruleId":"47","severity":1,"message":"48","line":9,"column":5,"nodeType":"49","endLine":11,"endColumn":45},"jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement"]
