--- src/components/HeaderBlock/index.js
@@ -1,14 +1,0 @@
-import './style.css';
-
-const HeaderBlock = ({ title }) => {
-	return (
-		<div>
-			<div>
-				<h1>{title}</h1>
-				<p>Simple Triple Triad Card Game</p>
-			</div>
-		</div>
-	);
-};
-
-export default HeaderBlock;
