--- src/components/HeaderBlock/style.css
@@ -1,3 +1,0 @@
-.header {
-	background-color: brown;
-}
