+++ src/components/Layout/index.js
@@ -1,6 +1,6 @@
 import styles from './style.module.css';
-import background from './assets/bg1.jpg';
-console.log(background);
+
+import cn from 'classnames';
 const Layout = ({ id, title, urlBg, colorBg, children }) => {
 	const style = {
 		background: colorBg,
@@ -15,9 +15,7 @@ const Layout = ({ id, title, urlBg, colorBg, children }) => {
 						<h3>{title}</h3>
 						<span className={styles.separator}></span>
 					</div>
-					<div className={`${styles.desc} ${styles.full}`}>
-						{children}
-					</div>
+					<div className={cn('full', 'desc')}>{children}</div>
 				</article>
 			</div>
 		</section>
