+++ src/components/PokemonCard/index.js
@@ -15,19 +15,19 @@ const PokemonCard = ({ name, img, id, type, values }) => {
 			<div
 				className={cn(s.pokemonCard, { [s.active]: isActive })}>
 				<div className={s.cardFront}>
-					<div className={`&{s.wrap} ${s.front}`}>
-						<div className={`${s.pokemon} ${s[type]}`}>
+					<div className={cn(s.wrap, s.front)}>
+						<div className={cn(s.pokemon, s[type])}>
 							<div className={s.value}>
-								<div className={`${s.count} ${s.top}`}>
+								<div className={cn(s.count, s.top)}>
 									{values.top}
 								</div>
-								<div className={`${s.count} ${s.right}`}>
+								<div className={cn(s.count, cn.right)}>
 									{values.right}
 								</div>
-								<div className={`${s.count} ${s.bottom}`}>
+								<div className={cn(s.count, s.bottom)}>
 									{values.bottom}
 								</div>
-								<div className={`${s.count} ${s.left}`}>
+								<div className={cn(s.count, s.left)}>
 									{values.left}
 								</div>
 							</div>
@@ -46,11 +46,8 @@ const PokemonCard = ({ name, img, id, type, values }) => {
 				</div>
 
 				<div className={s.cardBack}>
-					<div className={`${s.wrap} {s.back}`}>
-						<img
-							src={cardBackSide}
-							alt={`${s.Сard} ${s.Backed}`}
-						/>
+					<div className={cn(s.wrap, cn.back)}>
+						<img src={cardBackSide} alt={cn(s.Сard, s.Backed)} />
 					</div>
 				</div>
 			</div>
