+++ src/routes/Game/index.js
@@ -1,0 +1,5 @@
+const GamePage = () => {
+	return <div>dsfs</div>;
+};
+
+export default GamePage;
